refactor: split visualizer into different files
