+++ mermaid_visualizer_test.go
@@ -6,37 +6,6 @@ import (
 	"testing"
 )
 
-func TestGraphvizOutput(t *testing.T) {
-	fsmUnderTest := NewFSM(
-		"closed",
-		Events{
-			{Name: "open", Src: []string{"closed"}, Dst: "open"},
-			{Name: "close", Src: []string{"open"}, Dst: "closed"},
-			{Name: "part-close", Src: []string{"intermediate"}, Dst: "closed"},
-		},
-		Callbacks{},
-	)
-
-	got := Visualize(fsmUnderTest)
-	wanted := `
-digraph fsm {
-    "closed" -> "open" [ label = "open" ];
-    "intermediate" -> "closed" [ label = "part-close" ];
-    "open" -> "closed" [ label = "close" ];
-
-    "closed";
-    "intermediate";
-    "open";
-}`
-	normalizedGot := strings.ReplaceAll(got, "\n", "")
-	normalizedWanted := strings.ReplaceAll(wanted, "\n", "")
-	if normalizedGot != normalizedWanted {
-		t.Errorf("build graphivz graph failed. \nwanted \n%s\nand got \n%s\n", wanted, got)
-		fmt.Println([]byte(normalizedGot))
-		fmt.Println([]byte(normalizedWanted))
-	}
-}
-
 func TestMermaidOutput(t *testing.T) {
 	fsmUnderTest := NewFSM(
 		"closed",
@@ -48,7 +17,7 @@ func TestMermaidOutput(t *testing.T) {
 		Callbacks{},
 	)
 
-	got, err := VisualizeWithType(fsmUnderTest, MERMAID)
+	got, err := VisualizeForMermaidWithGraphType(fsmUnderTest, StateDiagram)
 	if err != nil {
 		t.Errorf("got error for visualizing with type MERMAID: %s", err)
 	}
@@ -67,7 +36,6 @@ stateDiagram
 		fmt.Println([]byte(normalizedWanted))
 	}
 }
-
 
 func TestMermaidFlowChartOutput(t *testing.T) {
 	fsmUnderTest := NewFSM(
